backend/requirements.txt
Django==5.0
Pillow==10.2.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-decouple==3.8
stripe
reportlab==4.0.9
Jinja2==3.1.2
